requirements.txt
click
nvidia-ml-py
PyYAML
